moving go code to src
