+++ src/file/file.go
@@ -1,4 +1,4 @@
-package lib
+package file
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 )
 
 func WriteToFile(version string, outputFilePath string) error {
-	data := []byte(fmt.Sprintf("BUILD_VERSION=%s", version))
+	data := fmt.Appendf([]byte{}, "BUILD_VERSION=%s", version)
 	err := os.WriteFile(outputFilePath, data, 0666)
 	if err != nil {
 		return err
