+++ src/file/file_test.go
@@ -1,4 +1,4 @@
-package lib
+package file
 
 import (
 	"fmt"
