+++ src/generator/generated_version_test.go
@@ -1,4 +1,4 @@
-package lib
+package generator
 
 import (
 	"strings"
