+++ src/git/git.go
@@ -1,4 +1,4 @@
-package lib
+package git
 
 import (
 	"regexp"
