+++ src/git/git_test.go
@@ -1,8 +1,9 @@
-package lib
+package git
 
 import (
 	"testing"
 
+	"github.com/alacrity-sg/build-version/src/lib"
 	bv_test "github.com/alacrity-sg/build-version/test"
 	"github.com/go-git/go-git/v5"
 )
@@ -13,9 +14,9 @@ func TestGetLatestReleaseTagSingle(t *testing.T) {
 	_, err := r.CreateTag(expectedTag, commit.Hash, &git.CreateTagOptions{
 		Message: expectedTag,
 	})
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	tag, err := GetLatestReleaseTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
@@ -27,14 +28,14 @@ func TestGetLatestReleaseTagMultiple(t *testing.T) {
 	_, err := r.CreateTag(unexpectedTag, commit.Hash, &git.CreateTagOptions{
 		Message: unexpectedTag,
 	})
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	expectedTag := "v1.0.1"
 	_, err = r.CreateTag(expectedTag, commit.Hash, &git.CreateTagOptions{
 		Message: expectedTag,
 	})
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	tag, err := GetLatestReleaseTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
@@ -46,14 +47,14 @@ func TestGetLatestReleaseTagMultipleWithRC(t *testing.T) {
 		Message: "Commit",
 	}
 	_, err := r.CreateTag("v1.0.0", commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	expectedTag := "v1.0.1"
 	_, err = r.CreateTag(expectedTag, commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	_, err = r.CreateTag("v1.0.0-rc.1234", commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	tag, err := GetLatestReleaseTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
@@ -66,9 +67,9 @@ func TestGetRCTagSingle(t *testing.T) {
 	}
 	expectedTag := "v1.0.0-rc.12345as"
 	_, err := r.CreateTag(expectedTag, commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	tag, err := GetLatestRCTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
@@ -80,12 +81,12 @@ func TestGetRCTagMultiple(t *testing.T) {
 		Message: "Commit",
 	}
 	_, err := r.CreateTag("v1.0.0-rc.0000", commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	expectedTag := "v1.0.0-rc.12345as"
 	_, err = r.CreateTag(expectedTag, commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	tag, err := GetLatestRCTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
@@ -97,13 +98,13 @@ func TestGetRCTagMultipleWithRelease(t *testing.T) {
 		Message: "Commit",
 	}
 	_, err := r.CreateTag("v1.0.0-rc.0000", commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	expectedTag := "v1.0.0-rc.12345as"
 	_, err = r.CreateTag(expectedTag, commit.Hash, tagOptions)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	_, err = r.CreateTag("v1.0.1", commit.Hash, tagOptions)
 	tag, err := GetLatestRCTag(dir)
-	CheckIfError(err)
+	lib.CheckIfError(err)
 	if *tag != expectedTag[1:] {
 		t.Fail()
 	}
